Quake3e networking, memory management, compiler warnings, fixes from last release
* quake3e server, snapshot, challenge, etc

* Update unix_shared.c

* remove update server + init challenge

* Update sv_client.c

* stateless challenge

* download code

* reconnectlimit not working

* remove autoupdate

* use common net_ip from q3e and pass adr by pointer

* Update WolfSources.cmake

* move unix sys_sleep to unix_main

* accurate ping

* fix rtcw unique tree entities missing from PVS

* add osp mode for snapped player positions

* fix gamestate overflow for increased rtcw msg size

* add reconnect timeout back

* reuse the nullcmd instead of resetting it

* use the memory management from quake3e

* fix compiler warnings for const variables

* quake3e downloads, packet handling

* fix compiler warnings for const variables

* fix compiler warnings on linux

+++ src/renderer_vk/tr_public.h
@@ -143,11 +143,11 @@ typedef struct {
 	// won't be freed
 	void ( *Hunk_Clear )( void );
 #ifdef HUNK_DEBUG
-	void    *( *Hunk_AllocDebug )( int size, ha_pref pref, char *label, char *file, int line );
+	void    *( *Hunk_AllocDebug )( size_t size, ha_pref pref, char *label, char *file, int line );
 #else
-	void    *( *Hunk_Alloc )( int size, ha_pref pref );
+	void    *( *Hunk_Alloc )( size_t size, ha_pref pref );
 #endif
-	void    *( *Hunk_AllocateTempMemory )( int size );
+	void	*(*Hunk_AllocateTempMemory)( size_t size );
 	void ( *Hunk_FreeTempMemory )( void *block );
 
 	// dynamic memory allocator for things that need to be freed
